refactor(settings): extract brand form sections from org-brand-context
Moves the self-contained overview/logos/fonts/colors form sections and
the auto-extract banner into their own file. They only read their own
props/form field, no shared closures with the page or the expandable
brand entry, so this is a pure relocation.
